Gen_AI_Transfer_Learning_Fine_Tuning_pipeline/requirements.txt
fastapi
uvicorn
transformers
datasets
torch
pillow
PyPDF2
python-docx
prometheus-client
streamlit